updated translation file and moved the function to pivot.styles.tsx
